split and move api definitions
