+++ pyDataverse/api/__init__.py
@@ -1,0 +1,6 @@
+from .api import Api  # noqa
+from .data_access import DataAccessApi  # noqa
+from .metrics import MetricsApi  # noqa
+from .native import NativeApi  # noqa
+from .search import SearchApi  # noqa
+from .sword import SwordApi  # noqa
