+++ pyDataverse/api/native.py
@@ -1,868 +1,17 @@
-"""Dataverse API wrapper for all it's API's."""
-
 import json
-from typing import Any, Dict, Optional
-import httpx
 import subprocess as sp
-from warnings import warn
 
-from httpx import ConnectError, Response
+from httpx import Response
+from pydantic import computed_field
 
-from pyDataverse.auth import ApiTokenAuth
+from pyDataverse.api.api import Api, DEPRECATION_GUARD
 from pyDataverse.exceptions import (
-    ApiAuthorizationError,
-    ApiUrlError,
-    DatasetNotFoundError,
-    DataverseNotEmptyError,
     DataverseNotFoundError,
     OperationFailedError,
+    ApiAuthorizationError,
+    DataverseNotEmptyError,
+    DatasetNotFoundError,
 )
-
-DEPRECATION_GUARD = object()
-
-
-class Api:
-    """Base class.
-
-    Parameters
-    ----------
-        base_url : str
-            Base URL of Dataverse instance. Without trailing `/` at the end.
-            e.g. `http://demo.dataverse.org`
-        api_token : str
-            Authentication token for the api.
-
-    Attributes
-    ----------
-    base_url
-    api_token
-    dataverse_version
-
-    """
-
-    def __init__(
-        self,
-        base_url: str,
-        api_token: Optional[str] = None,
-        api_version: str = "latest",
-        *,
-        auth: Optional[httpx.Auth] = None,
-    ):
-        """Init an Api() class.
-
-        Scheme, host and path combined create the base-url for the api.
-        See more about URL at `Wikipedia <https://en.wikipedia.org/wiki/URL>`_.
-
-        Parameters
-        ----------
-        base_url : str
-            Base url for Dataverse api.
-        api_token : str | None
-            API token for Dataverse API. If you provide an :code:`api_token`, we
-            assume it is an API token as retrieved via your Dataverse instance
-            user profile.
-            We recommend using the :code:`auth` argument instead.
-            To retain the current behaviour with the :code:`auth` argument, change
-
-            .. code-block:: python
-
-                Api("https://demo.dataverse.org", "my_token")
-
-            to
-
-            .. code-block:: python
-
-                from pyDataverse.auth import ApiTokenAuth
-
-                Api("https://demo.dataverse.org", auth=ApiTokenAuth("my_token"))
-
-            If you are using an OIDC/OAuth 2.0 Bearer token, please use the :code:`auth`
-            parameter with the :py:class:`.auth.BearerTokenAuth`.
-        api_version : str
-            The version string of the Dataverse API or :code:`latest`, e.g.,
-            :code:`v1`. Defaults to :code:`latest`, which drops the version from
-            the API urls.
-        auth : httpx.Auth | None
-            You can provide any authentication mechanism you like to connect to
-            your Dataverse instance.  The most common mechanisms are implemented
-            in :py:mod:`.auth`, but if one is missing, you can use your own
-            `httpx.Auth`-compatible class. For more information, have a look at
-            `httpx' Authentication docs
-            <https://www.python-httpx.org/advanced/authentication/>`_.
-        Examples
-        -------
-        Create an API connection::
-
-        .. code-block::
-
-            >>> from pyDataverse.api import Api
-            >>> base_url = 'http://demo.dataverse.org'
-            >>> api = Api(base_url)
-
-        .. code-block::
-
-            >>> from pyDataverse.api import Api
-            >>> from pyDataverse.auth import ApiTokenAuth
-            >>> base_url = 'http://demo.dataverse.org'
-            >>> api = Api(base_url, ApiTokenAuth('my_api_token'))
-
-        """
-        if not isinstance(base_url, str):
-            raise ApiUrlError("base_url {0} is not a string.".format(base_url))
-
-        self.base_url = base_url.rstrip("/")
-        self.client = None
-
-        if not isinstance(api_version, str):
-            raise ApiUrlError("api_version {0} is not a string.".format(api_version))
-        self.api_version = api_version
-
-        self.auth = auth
-        self.api_token = api_token
-        if api_token is not None:
-            if auth is None:
-                self.auth = ApiTokenAuth(api_token)
-            else:
-                self.api_token = None
-                warn(
-                    UserWarning(
-                        "You provided both, an api_token and a custom auth "
-                        "method. We will only use the auth method."
-                    )
-                )
-
-        if self.api_version == "latest":
-            self.base_url_api = "{0}/api".format(self.base_url)
-        else:
-            self.base_url_api = "{0}/api/{1}".format(self.base_url, self.api_version)
-
-        self.timeout = 500
-
-    def __str__(self):
-        """Return the class name and URL of the used API class.
-
-        Returns
-        -------
-        str
-            Naming of the API class.
-
-        """
-        return f"{self.__class__.__name__}: {self.base_url_api}"
-
-    def get_request(self, url, params=None, auth=DEPRECATION_GUARD):
-        """Make a GET request.
-
-        Parameters
-        ----------
-        url : str
-            Full URL.
-        params : dict
-            Dictionary of parameters to be passed with the request.
-            Defaults to `None`.
-        auth : Any
-            .. deprecated:: 0.3.4
-                The auth parameter was ignored before version 0.3.4.
-                Please pass your auth to the Api instance directly, as
-                explained in :py:func:`Api.__init__`.
-                If you need multiple auth methods, create multiple
-                API instances:
-
-                .. code-block:: python
-
-                    api = Api("https://demo.dataverse.org", auth=ApiTokenAuth("my_api_token"))
-                    api_oauth = Api("https://demo.dataverse.org", auth=BearerTokenAuth("my_bearer_token"))
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        if auth is not DEPRECATION_GUARD:
-            warn(
-                DeprecationWarning(
-                    "The auth parameter is deprecated. Please pass your auth "
-                    "arguments to the __init__ method instead."
-                )
-            )
-        headers = {}
-        headers["User-Agent"] = "pydataverse"
-
-        if self.client is None:
-            return self._sync_request(
-                method=httpx.get,
-                url=url,
-                headers=headers,
-                params=params,
-            )
-        else:
-            return self._async_request(
-                method=self.client.get,
-                url=url,
-                headers=headers,
-                params=params,
-            )
-
-    def post_request(
-        self, url, data=None, auth=DEPRECATION_GUARD, params=None, files=None
-    ):
-        """Make a POST request.
-
-        params will be added as key-value pairs to the URL.
-
-        Parameters
-        ----------
-        url : str
-            Full URL.
-        data : str
-            Metadata as a json-formatted string. Defaults to `None`.
-        auth : Any
-            .. deprecated:: 0.3.4
-                The auth parameter was ignored before version 0.3.4.
-                Please pass your auth to the Api instance directly, as
-                explained in :py:func:`Api.__init__`.
-                If you need multiple auth methods, create multiple
-                API instances:
-
-                .. code-block:: python
-
-                    api = Api("https://demo.dataverse.org", auth=ApiTokenAuth("my_api_token"))
-                    api_oauth = Api("https://demo.dataverse.org", auth=BearerTokenAuth("my_bearer_token"))
-        files : dict
-            e.g. :code:`files={'file': open('sample_file.txt','rb')}`
-        params : dict
-            Dictionary of parameters to be passed with the request.
-            Defaults to :code:`None`.
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        if auth is not DEPRECATION_GUARD:
-            warn(
-                DeprecationWarning(
-                    "The auth parameter is deprecated. Please pass your auth "
-                    "arguments to the __init__ method instead."
-                )
-            )
-        headers = {}
-        headers["User-Agent"] = "pydataverse"
-
-        if isinstance(data, str):
-            data = json.loads(data)
-
-        # Decide whether to use 'data' or 'json' args
-        request_params = self._check_json_data_form(data)
-
-        if self.client is None:
-            return self._sync_request(
-                method=httpx.post,
-                url=url,
-                headers=headers,
-                params=params,
-                files=files,
-                **request_params,
-            )
-        else:
-            return self._async_request(
-                method=self.client.post,
-                url=url,
-                headers=headers,
-                params=params,
-                files=files,
-                **request_params,
-            )
-
-    def put_request(self, url, data=None, auth=DEPRECATION_GUARD, params=None):
-        """Make a PUT request.
-
-        Parameters
-        ----------
-        url : str
-            Full URL.
-        data : str
-            Metadata as a json-formatted string. Defaults to `None`.
-        auth : Any
-            .. deprecated:: 0.3.4
-                The auth parameter was ignored before version 0.3.4.
-                Please pass your auth to the Api instance directly, as
-                explained in :py:func:`Api.__init__`.
-                If you need multiple auth methods, create multiple
-                API instances:
-
-                .. code-block:: python
-
-                    api = Api("https://demo.dataverse.org", auth=ApiTokenAuth("my_api_token"))
-                    api_oauth = Api("https://demo.dataverse.org", auth=BearerTokenAuth("my_bearer_token"))
-        params : dict
-            Dictionary of parameters to be passed with the request.
-            Defaults to `None`.
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        if auth is not DEPRECATION_GUARD:
-            warn(
-                DeprecationWarning(
-                    "The auth parameter is deprecated. Please pass your auth "
-                    "arguments to the __init__ method instead."
-                )
-            )
-        headers = {}
-        headers["User-Agent"] = "pydataverse"
-
-        if isinstance(data, str):
-            data = json.loads(data)
-
-        # Decide whether to use 'data' or 'json' args
-        request_params = self._check_json_data_form(data)
-
-        if self.client is None:
-            return self._sync_request(
-                method=httpx.put,
-                url=url,
-                json=data,
-                headers=headers,
-                params=params,
-                **request_params,
-            )
-        else:
-            return self._async_request(
-                method=self.client.put,
-                url=url,
-                json=data,
-                headers=headers,
-                params=params,
-                **request_params,
-            )
-
-    def delete_request(self, url, auth=DEPRECATION_GUARD, params=None):
-        """Make a Delete request.
-
-        Parameters
-        ----------
-        url : str
-            Full URL.
-        auth : Any
-            .. deprecated:: 0.3.4
-                The auth parameter was ignored before version 0.3.4.
-                Please pass your auth to the Api instance directly, as
-                explained in :py:func:`Api.__init__`.
-                If you need multiple auth methods, create multiple
-                API instances:
-
-                .. code-block:: python
-
-                    api = Api("https://demo.dataverse.org", auth=ApiTokenAuth("my_api_token"))
-                    api_oauth = Api("https://demo.dataverse.org", auth=BearerTokenAuth("my_bearer_token"))
-        params : dict
-            Dictionary of parameters to be passed with the request.
-            Defaults to `None`.
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        if auth is not DEPRECATION_GUARD:
-            warn(
-                DeprecationWarning(
-                    "The auth parameter is deprecated. Please pass your auth "
-                    "arguments to the __init__ method instead."
-                )
-            )
-        headers = {}
-        headers["User-Agent"] = "pydataverse"
-
-        if self.client is None:
-            return self._sync_request(
-                method=httpx.delete,
-                url=url,
-                headers=headers,
-                params=params,
-            )
-        else:
-            return self._async_request(
-                method=self.client.delete,
-                url=url,
-                headers=headers,
-                params=params,
-            )
-
-    @staticmethod
-    def _check_json_data_form(data: Optional[Dict]):
-        """This method checks and distributes given payload to match Dataverse expectations.
-
-        In the case of the form-data keyed by "jsonData", Dataverse expects
-        the payload as a string in a form of a dictionary. This is not possible
-        using HTTPXs json parameter, so we need to handle this case separately.
-        """
-
-        if not data:
-            return {}
-        elif not isinstance(data, dict):
-            raise ValueError("Data must be a dictionary.")
-        elif "jsonData" not in data:
-            return {"json": data}
-
-        assert list(data.keys()) == [
-            "jsonData"
-        ], "jsonData must be the only key in the dictionary."
-
-        # Content of JSON data should ideally be a string
-        content = data["jsonData"]
-        if not isinstance(content, str):
-            data["jsonData"] = json.dumps(content)
-
-        return {"data": data}
-
-    def _sync_request(
-        self,
-        method,
-        **kwargs,
-    ):
-        """
-        Sends a synchronous request to the specified URL using the specified HTTP method.
-
-        Args:
-            method (function): The HTTP method to use for the request.
-            **kwargs: Additional keyword arguments to be passed to the method.
-
-        Returns:
-            httpx.Response: The response object returned by the request.
-
-        Raises:
-            ApiAuthorizationError: If the response status code is 401 (Authorization error).
-            ConnectError: If a connection to the API cannot be established.
-        """
-        assert "url" in kwargs, "URL is required for a request."
-
-        kwargs = self._filter_kwargs(kwargs)
-
-        try:
-            resp: httpx.Response = method(
-                **kwargs, auth=self.auth, follow_redirects=True, timeout=None
-            )
-            if resp.status_code == 401:
-                try:
-                    error_msg = resp.json()["message"]
-                except json.JSONDecodeError:
-                    error_msg = resp.reason_phrase
-                raise ApiAuthorizationError(
-                    "ERROR: HTTP 401 - Authorization error {0}. MSG: {1}".format(
-                        kwargs["url"], error_msg
-                    )
-                )
-
-            return resp
-
-        except ConnectError:
-            raise ConnectError(
-                "ERROR - Could not establish connection to api '{0}'.".format(
-                    kwargs["url"]
-                )
-            )
-
-    async def _async_request(
-        self,
-        method,
-        **kwargs,
-    ):
-        """
-        Sends an asynchronous request to the specified URL using the specified HTTP method.
-
-        Args:
-            method (callable): The HTTP method to use for the request.
-            **kwargs: Additional keyword arguments to be passed to the method.
-
-        Raises:
-            ApiAuthorizationError: If the response status code is 401 (Authorization error).
-            ConnectError: If a connection to the API cannot be established.
-
-        Returns:
-            The response object.
-
-        """
-        assert "url" in kwargs, "URL is required for a request."
-
-        kwargs = self._filter_kwargs(kwargs)
-
-        try:
-            resp = await method(**kwargs, auth=self.auth)
-
-            if resp.status_code == 401:
-                error_msg = resp.json()["message"]
-                raise ApiAuthorizationError(
-                    "ERROR: HTTP 401 - Authorization error {0}. MSG: {1}".format(
-                        kwargs["url"], error_msg
-                    )
-                )
-
-            return resp
-
-        except ConnectError:
-            raise ConnectError(
-                "ERROR - Could not establish connection to api '{0}'.".format(
-                    kwargs["url"]
-                )
-            )
-
-    @staticmethod
-    def _filter_kwargs(kwargs: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Filters out any keyword arguments that are `None` from the specified dictionary.
-
-        Args:
-            kwargs (Dict[str, Any]): The dictionary to filter.
-
-        Returns:
-            Dict[str, Any]: The filtered dictionary.
-        """
-        return {k: v for k, v in kwargs.items() if v is not None}
-
-    async def __aenter__(self):
-        """
-        Context manager method that initializes an instance of httpx.AsyncClient.
-
-        Returns:
-            httpx.AsyncClient: An instance of httpx.AsyncClient.
-        """
-        self.client = httpx.AsyncClient()
-
-    async def __aexit__(self, exc_type, exc_value, traceback):
-        """
-        Closes the client connection when exiting a context manager.
-
-        Args:
-            exc_type (type): The type of the exception raised, if any.
-            exc_value (Exception): The exception raised, if any.
-            traceback (traceback): The traceback object associated with the exception, if any.
-        """
-
-        await self.client.aclose()
-        self.client = None
-
-
-class DataAccessApi(Api):
-    """Class to access Dataverse's Data Access API.
-
-    Attributes
-    ----------
-    base_url_api_data_access : type
-        Description of attribute `base_url_api_data_access`.
-    base_url : type
-        Description of attribute `base_url`.
-
-    """
-
-    def __init__(self, base_url, api_token=None, *, auth=None):
-        """Init an DataAccessApi() class."""
-        super().__init__(base_url, api_token, auth=auth)
-        if base_url:
-            self.base_url_api_data_access = "{0}/access".format(self.base_url_api)
-        else:
-            self.base_url_api_data_access = self.base_url_api
-
-    def get_datafile(
-        self,
-        identifier,
-        data_format=None,
-        no_var_header=None,
-        image_thumb=None,
-        is_pid=True,
-        auth=DEPRECATION_GUARD,
-    ):
-        """Download a datafile via the Dataverse Data Access API.
-
-        Get by file id (HTTP Request).
-
-        .. code-block:: bash
-
-            GET /api/access/datafile/$id
-
-        Get by persistent identifier (HTTP Request).
-
-        .. code-block:: bash
-
-            GET http://$SERVER/api/access/datafile/:persistentId/?persistentId=doi:10.5072/FK2/J8SJZB
-
-        Parameters
-        ----------
-        identifier : str
-            Identifier of the datafile. Can be datafile id or persistent
-            identifier of the datafile (e. g. doi).
-        is_pid : bool
-            ``True`` to use persistent identifier. ``False``, if not.
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        is_first_param = True
-        if is_pid:
-            url = "{0}/datafile/:persistentId/?persistentId={1}".format(
-                self.base_url_api_data_access, identifier
-            )
-        else:
-            url = "{0}/datafile/{1}".format(self.base_url_api_data_access, identifier)
-            if data_format or no_var_header or image_thumb:
-                url += "?"
-        if data_format:
-            url += "format={0}".format(data_format)
-            is_first_param = False
-        if no_var_header:
-            if not is_first_param:
-                url += "&"
-            url += "noVarHeader={0}".format(no_var_header)
-            is_first_param = False
-        if image_thumb:
-            if not is_first_param:
-                url += "&"
-            url += "imageThumb={0}".format(image_thumb)
-        return self.get_request(url, auth=auth)
-
-    def get_datafiles(self, identifier, data_format=None, auth=DEPRECATION_GUARD):
-        """Download a datafile via the Dataverse Data Access API.
-
-        Get by file id (HTTP Request).
-
-        .. code-block:: bash
-
-            GET /api/access/datafiles/$id1,$id2,...$idN
-
-        Get by persistent identifier (HTTP Request).
-
-        Parameters
-        ----------
-        identifier : str
-            Identifier of the dataset. Can be datafile id or persistent
-            identifier of the datafile (e. g. doi).
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        url = "{0}/datafiles/{1}".format(self.base_url_api_data_access, identifier)
-        if data_format:
-            url += "?format={0}".format(data_format)
-        return self.get_request(url, auth=auth)
-
-    def get_datafile_bundle(
-        self, identifier, file_metadata_id=None, auth=DEPRECATION_GUARD
-    ):
-        """Download a datafile in all its formats.
-
-        HTTP Request:
-
-        .. code-block:: bash
-
-            GET /api/access/datafile/bundle/$id
-
-        Data Access API calls can now be made using persistent identifiers (in
-        addition to database ids). This is done by passing the constant
-        :persistentId where the numeric id of the file is expected, and then
-        passing the actual persistent id as a query parameter with the name
-        persistentId.
-
-        This is a convenience packaging method available for tabular data
-        files. It returns a zipped bundle that contains the data in the
-        following formats:
-        - Tab-delimited;
-        - “Saved Original”, the proprietary (SPSS, Stata, R, etc.) file
-        from which the tabular data was ingested;
-        - Generated R Data frame (unless the “original” above was in R);
-        - Data (Variable) metadata record, in DDI XML;
-        - File citation, in Endnote and RIS formats.
-
-        Parameters
-        ----------
-        identifier : str
-            Identifier of the dataset.
-
-        Returns
-        -------
-        httpx.Response
-            Response object of httpx library.
-
-        """
-        url = "{0}/datafile/bundle/{1}".format(
-            self.base_url_api_data_access, identifier
-        )
-        if file_metadata_id:
-            url += "?fileMetadataId={0}".format(file_metadata_id)
-        return self.get_request(url, auth=auth)
-
-    def request_access(self, identifier, auth=DEPRECATION_GUARD, is_filepid=False):
-        """Request datafile access.
-
-        This method requests access to the datafile whose id is passed on the behalf of an authenticated user whose key is passed. Note that not all datasets allow access requests to restricted files.
-
-        https://guides.dataverse.org/en/4.18.1/api/dataaccess.html#request-access
-
-        /api/access/datafile/$id/requestAccess
-
-        curl -H "X-Dataverse-key:$API_TOKEN" -X PUT http://$SERVER/api/access/datafile/{id}/requestAccess
-        """
-        if is_filepid:
-            url = "{0}/datafile/:persistentId/requestAccess?persistentId={1}".format(
-                self.base_url_api_data_access, identifier
-            )
-        else:
-            url = "{0}/datafile/{1}/requestAccess".format(
-                self.base_url_api_data_access, identifier
-            )
-        return self.put_request(url, auth=auth)
-
-    def allow_access_request(
-        self, identifier, do_allow=True, auth=DEPRECATION_GUARD, is_pid=True
-    ):
-        """Allow access request for datafiles.
-
-        https://guides.dataverse.org/en/latest/api/dataaccess.html#allow-access-requests
-
-        curl -H "X-Dataverse-key:$API_TOKEN" -X PUT -d true http://$SERVER/api/access/{id}/allowAccessRequest
-        curl -H "X-Dataverse-key:$API_TOKEN" -X PUT -d true http://$SERVER/api/access/:persistentId/allowAccessRequest?persistentId={pid}
-        """
-        if is_pid:
-            url = "{0}/:persistentId/allowAccessRequest?persistentId={1}".format(
-                self.base_url_api_data_access, identifier
-            )
-        else:
-            url = "{0}/{1}/allowAccessRequest".format(
-                self.base_url_api_data_access, identifier
-            )
-
-        if do_allow:
-            data = "true"
-        else:
-            data = "false"
-        return self.put_request(url, data=data, auth=auth)
-
-    def grant_file_access(self, identifier, user, auth=DEPRECATION_GUARD):
-        """Grant datafile access.
-
-        https://guides.dataverse.org/en/4.18.1/api/dataaccess.html#grant-file-access
-
-        curl -H "X-Dataverse-key:$API_TOKEN" -X PUT http://$SERVER/api/access/datafile/{id}/grantAccess/{@userIdentifier}
-        """
-        url = "{0}/datafile/{1}/grantAccess/{2}".format(
-            self.base_url_api_data_access, identifier, user
-        )
-        return self.put_request(url, auth=auth)
-
-    def list_file_access_requests(self, identifier, auth=DEPRECATION_GUARD):
-        """Liste datafile access requests.
-
-        https://guides.dataverse.org/en/4.18.1/api/dataaccess.html#list-file-access-requests
-
-        curl -H "X-Dataverse-key:$API_TOKEN" -X GET http://$SERVER/api/access/datafile/{id}/listRequests
-        """
-        url = "{0}/datafile/{1}/listRequests".format(
-            self.base_url_api_data_access, identifier
-        )
-        return self.get_request(url, auth=auth)
-
-
-class MetricsApi(Api):
-    """Class to access Dataverse's Metrics API.
-
-    Attributes
-    ----------
-    base_url_api_metrics : type
-        Description of attribute `base_url_api_metrics`.
-    base_url : type
-        Description of attribute `base_url`.
-
-    """
-
-    def __init__(self, base_url, api_token=None, api_version="latest", *, auth=None):
-        """Init an MetricsApi() class."""
-        super().__init__(base_url, api_token, api_version, auth=auth)
-        if base_url:
-            self.base_url_api_metrics = "{0}/api/info/metrics".format(self.base_url)
-        else:
-            self.base_url_api_metrics = None
-
-    def total(self, data_type, date_str=None, auth=DEPRECATION_GUARD):
-        """
-        GET https://$SERVER/api/info/metrics/$type
-        GET https://$SERVER/api/info/metrics/$type/toMonth/$YYYY-DD
-
-        $type can be set to dataverses, datasets, files or downloads.
-
-        """
-        url = "{0}/{1}".format(self.base_url_api_metrics, data_type)
-        if date_str:
-            url += "/toMonth/{0}".format(date_str)
-        return self.get_request(url, auth=auth)
-
-    def past_days(self, data_type, days_str, auth=DEPRECATION_GUARD):
-        """
-
-        http://guides.dataverse.org/en/4.18.1/api/metrics.html
-        GET https://$SERVER/api/info/metrics/$type/pastDays/$days
-
-        $type can be set to dataverses, datasets, files or downloads.
-        """
-        # TODO: check if date-string has proper format
-        url = "{0}/{1}/pastDays/{2}".format(
-            self.base_url_api_metrics, data_type, days_str
-        )
-        return self.get_request(url, auth=auth)
-
-    def get_dataverses_by_subject(self, auth=DEPRECATION_GUARD):
-        """
-        GET https://$SERVER/api/info/metrics/dataverses/bySubject
-
-        $type can be set to dataverses, datasets, files or downloads.
-        """
-        # TODO: check if date-string has proper format
-        url = "{0}/dataverses/bySubject".format(self.base_url_api_metrics)
-        return self.get_request(url, auth=auth)
-
-    def get_dataverses_by_category(self, auth=DEPRECATION_GUARD):
-        """
-        GET https://$SERVER/api/info/metrics/dataverses/byCategory
-
-        $type can be set to dataverses, datasets, files or downloads.
-        """
-        # TODO: check if date-string has proper format
-        url = "{0}/dataverses/byCategory".format(self.base_url_api_metrics)
-        return self.get_request(url, auth=auth)
-
-    def get_datasets_by_subject(self, date_str=None, auth=DEPRECATION_GUARD):
-        """
-        GET https://$SERVER/api/info/metrics/datasets/bySubject
-
-        $type can be set to dataverses, datasets, files or downloads.
-        """
-        # TODO: check if date-string has proper format
-        url = "{0}/datasets/bySubject".format(self.base_url_api_metrics)
-        if date_str:
-            url += "/toMonth/{0}".format(date_str)
-        return self.get_request(url, auth=auth)
-
-    def get_datasets_by_data_location(self, data_location, auth=DEPRECATION_GUARD):
-        """
-        GET https://$SERVER/api/info/metrics/datasets/?dataLocation=$location
-
-        $type can be set to dataverses, datasets, files or downloads.
-        """
-        # TODO: check if date-string has proper format
-        url = "{0}/datasets/?dataLocation={1}".format(
-            self.base_url_api_metrics, data_location
-        )
-        return self.get_request(url, auth=auth)
 
 
 class NativeApi(Api):
@@ -886,20 +35,9 @@ class NativeApi(Api):
 
     """
 
-    def __init__(self, base_url: str, api_token=None, api_version="v1", *, auth=None):
-        """Init an Api() class.
-
-        Scheme, host and path combined create the base-url for the api.
-        See more about URL at `Wikipedia <https://en.wikipedia.org/wiki/URL>`_.
-
-        Parameters
-        ----------
-        native_api_version : str
-            API version of Dataverse native API. Default is `v1`.
-
-        """
-        super().__init__(base_url, api_token, api_version, auth=auth)
-        self.base_url_api_native = self.base_url_api
+    @computed_field(return_type=str)
+    def base_url_api_native(self):
+        return self.base_url_api
 
     def get_dataverse(self, identifier, auth=DEPRECATION_GUARD):
         """Get dataverse metadata by alias or id.
@@ -2540,152 +1678,3 @@ class NativeApi(Api):
         else:
             url = f"{self.base_url_api_native}/files/{identifier}/restrict"
         return self.put_request(url, auth=True)
-
-
-class SearchApi(Api):
-    """Class to access Dataverse's Search API.
-
-    Examples
-    -------
-    Examples should be written in doctest format, and
-    should illustrate how to use the function/class.
-    >>>
-
-    Attributes
-    ----------
-    base_url_api_search : type
-        Description of attribute `base_url_api_search`.
-    base_url : type
-        Description of attribute `base_url`.
-
-    """
-
-    def __init__(self, base_url, api_token=None, api_version="latest", *, auth=None):
-        """Init an SearchApi() class."""
-        super().__init__(base_url, api_token, api_version, auth=auth)
-        if base_url:
-            self.base_url_api_search = "{0}/search?q=".format(self.base_url_api)
-        else:
-            self.base_url_api_search = self.base_url_api
-
-    def search(
-        self,
-        q_str,
-        data_type=None,
-        subtree=None,
-        sort=None,
-        order=None,
-        per_page=None,
-        start=None,
-        show_relevance=None,
-        show_facets=None,
-        filter_query=None,
-        show_entity_ids=None,
-        query_entities=None,
-        auth=DEPRECATION_GUARD,
-    ):
-        """Search.
-
-        http://guides.dataverse.org/en/4.18.1/api/search.html
-        """
-        url = "{0}{1}".format(self.base_url_api_search, q_str)
-        if data_type:
-            # TODO: pass list of types
-            url += "&type={0}".format(data_type)
-        if subtree:
-            # TODO: pass list of subtrees
-            url += "&subtree={0}".format(subtree)
-        if sort:
-            url += "&sort={0}".format(sort)
-        if order:
-            url += "&order={0}".format(order)
-        if per_page:
-            url += "&per_page={0}".format(per_page)
-        if start:
-            url += "&start={0}".format(start)
-        if show_relevance:
-            url += "&show_relevance={0}".format(show_relevance)
-        if show_facets:
-            url += "&show_facets={0}".format(show_facets)
-        if filter_query:
-            url += "&fq={0}".format(filter_query)
-        if show_entity_ids:
-            url += "&show_entity_ids={0}".format(show_entity_ids)
-        if query_entities:
-            url += "&query_entities={0}".format(query_entities)
-        return self.get_request(url, auth=auth)
-
-
-class SwordApi(Api):
-    """Class to access Dataverse's SWORD API.
-
-    Parameters
-    ----------
-    sword_api_version : str
-        SWORD API version. Defaults to 'v1.1'.
-
-    Attributes
-    ----------
-    base_url_api_sword : str
-        Description of attribute `base_url_api_sword`.
-    base_url : str
-        Description of attribute `base_url`.
-    native_api_version : str
-        Description of attribute `native_api_version`.
-    sword_api_version
-
-    """
-
-    def __init__(
-        self,
-        base_url,
-        api_version="v1.1",
-        api_token=None,
-        sword_api_version="v1.1",
-        *,
-        auth=None,
-    ):
-        """Init a :class:`SwordApi <pyDataverse.api.SwordApi>` instance.
-
-        Parameters
-        ----------
-        sword_api_version : str
-            Api version of Dataverse SWORD API.
-        api_token : str | None
-            An Api token as retrieved from your Dataverse instance.
-        auth : httpx.Auth
-            Note that the SWORD API uses a different authentication mechanism
-            than the native API, in particular it uses `HTTP Basic
-            Authentication
-            <https://guides.dataverse.org/en/latest/api/sword.html#sword-auth>`_.
-            Thus, if you pass an api_token, it will be used as the username in
-            the HTTP Basic Authentication. If you pass a custom :py:class:`httpx.Auth`, use
-            :py:class:`httpx.BasicAuth` with an empty password:
-
-            .. code-block:: python
-
-                sword_api = Api(
-                    "https://demo.dataverse.org", auth=httpx.BasicAuth(username="my_token", password="")
-                )
-
-        """
-        if auth is None and api_token is not None:
-            auth = httpx.BasicAuth(api_token, "")
-        super().__init__(base_url, api_token, api_version, auth=auth)
-        if not isinstance(sword_api_version, ("".__class__, "".__class__)):
-            raise ApiUrlError(
-                "sword_api_version {0} is not a string.".format(sword_api_version)
-            )
-        self.sword_api_version = sword_api_version
-
-        # Test connection.
-        if self.base_url and sword_api_version:
-            self.base_url_api_sword = "{0}/dvn/api/data-deposit/{1}".format(
-                self.base_url, self.sword_api_version
-            )
-        else:
-            self.base_url_api_sword = base_url
-
-    def get_service_document(self):
-        url = "{0}/swordv2/service-document".format(self.base_url_api_sword)
-        return self.get_request(url, auth=True)
